Use the context render to check application notice
