+++ base/views/common.py
@@ -38,6 +38,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from base import models as mdl
 from base.models.utils import native
+from osis_common.models import application_notice
 from . import layout
 
 ITEMS_PER_PAGE = 25
@@ -83,10 +84,27 @@ def common_context_processor(request):
     else:
         sentry_dns = ''
     release_tag = settings.RELEASE_TAG if hasattr(settings, 'RELEASE_TAG') else None
-    return {'installed_apps': settings.INSTALLED_APPS,
-            'environment': env,
-            'sentry_dns': sentry_dns,
-            'release_tag': release_tag}
+
+    context = {'installed_apps': settings.INSTALLED_APPS,
+               'environment': env,
+               'sentry_dns': sentry_dns,
+               'release_tag': release_tag}
+
+    _check_notice(request, context)
+    return context
+
+
+def _check_notice(request, values):
+    if 'subject' not in request.session and 'notice' not in request.session:
+        notice = application_notice.find_current_notice()
+        if notice:
+            request.session.set_expiry(3600)
+            request.session['subject'] = notice.subject
+            request.session['notice'] = notice.notice
+
+    if 'subject' in request.session and 'notice' in request.session:
+        values['subject'] = request.session['subject']
+        values['notice'] = request.session['notice']
 
 
 def login(request):
