+++ base/views/layout.py
@@ -24,30 +24,15 @@
 #
 ##############################################################################
 from django import shortcuts
-from random import randint
-from osis_common.models import application_notice
+
+from osis_common.decorators.deprecated import deprecated
 
 
-def _check_notice(request, values):
-    if 'subject' not in request.session and 'notice' not in request.session:
-        notice = application_notice.find_current_notice()
-        if notice:
-            request.session.set_expiry(3600)
-            request.session['subject'] = notice.subject
-            request.session['notice'] = notice.notice
-
-    if 'subject' in request.session and 'notice' in request.session:
-        values['subject'] = request.session['subject']
-        values['notice'] = request.session['notice']
-
-
+@deprecated
 def render(request, template, values):
-    _check_notice(request, values)
-
     return shortcuts.render(request, template, values)
 
 
+@deprecated
 def render_to_response(request, template, values):
-    _check_notice(request, values)
-
     return shortcuts.render_to_response(template, values)
